# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Udemy API'
  description: 'BY Phan Văn Tuấn '
  contact:
    name: 'Swagger API Team'
  version: '1.0'
servers:
  -
    url: 'http://localhost:8000'
    description: 'API server'
  -
    url: 'https://api-study.salyr.online'
    description: 'production server'
paths:
...
  /api/courses:
    get:
      tags:
        - Courses
      summary: 'Get a paginated listing of courses'
      description: 'Display a paginated listing of courses.'
      operationId: 02eeda67042dba2d22a517f26fb8eaf1
      parameters:
        -
          name: pageNum
          in: query
          description: 'Page number'
          required: false
          schema:
            type: integer
        -
          name: pageSize
          in: query
          description: 'Number of items per page'
          required: false
          schema:
            type: integer
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseResource'
    post:
      tags:
        - Courses
      summary: 'Store a newly created course'
      description: 'Store a newly created course in storage.'
      operationId: a4d6b7e7be59571034428ffbb86aaf0a
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StoreCourseRequest'
      responses:
        201:
          description: 'Course created successfully'
          content:
            application/json:
              schema: []
      security:
        -
          bearerAuth: []
  '/api/courses/{course_id}':
    get:
      tags:
        - Courses
      summary: 'Display the specified course'
      description: 'Display the specified course.'
      operationId: e60bc21136ac1e1be61ca244ec552eae
      parameters:
        -
          name: course_id
          in: path
          description: 'ID of the course'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResource'
        404:
          description: 'Course not found'
    post:
      tags:
        - Courses
      summary: 'Update the specified course'
      description: 'Update the specified course in storage.'
      operationId: 3cef984cfe9942b9eafb4096512ad771
      parameters:
        -
          name: course_id
          in: path
          description: 'ID of the course'
          required: true
          schema:
            type: integer
        -
          name: _method
          in: query
          description: 'Method to be used'
          required: false
          schema:
            type: string
            default: PUT
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateCourseRequest'
      responses:
        200:
          description: 'Course updated successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResource'
        404:
          description: 'Course not found'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Courses
      summary: 'Remove the specified course'
      description: 'Remove the specified course from storage.'
      operationId: bec73dfbe0ae9a5de705f020d589ded1
      parameters:
        -
          name: course_id
          in: path
          description: 'ID of the course'
          required: true
          schema:
            type: integer
      responses:
        204:
          description: 'Course deleted successfully'
        404:
          description: 'Course not found'
      security:
        -
          bearerAuth: []
  '/api/courses/{course_id}/reviews':
    get:
      tags:
        - Courses
      summary: 'Display a listing of reviews for the specified course'
      description: 'Display a listing of reviews for the specified course.'
      operationId: dc9ee6917a87f94e8c044fa90a62bdc6
      parameters:
        -
          name: course_id
          in: path
          description: 'ID of the course'
          required: true
          schema:
            type: integer
        -
          name: pageNum
          in: query
          description: 'Page number'
          required: false
          schema:
            type: integer
        -
          name: pageSize
          in: query
          description: 'Number of items per page'
          required: false
          schema:
            type: integer
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseReviewResource'
        404:
          description: 'Course not found'
  '/api/courses/{course_id}/chapters':
    get:
      tags:
        - Courses
      summary: 'Display a listing of chapters for the specified course'
      description: 'Display a listing of chapters for the specified course.'
      operationId: 0cc05de83178774f4f095b6fb21648e8
      parameters:
        -
          name: course_id
          in: path
          description: 'ID of the course'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChapterResource'
        404:
          description: 'Course not found'
  /api/user/course-reviews:
    get:
      tags:
        - CourseReviews
      summary: 'Get a list of course reviews'
      operationId: 8c3c1b2ec9e8735c37ff191c457ea006
      parameters:
        -
          name: pageNum
          in: query
          description: 'Page number'
          required: false
          schema:
            type: integer
            example: 1
        -
          name: pageSize
          in: query
          description: 'Number of items per page'
          required: false
          schema:
            type: integer
            example: 15
      responses:
        200:
          description: 'List of course reviews'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseReviewResource'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - CourseReviews
      summary: 'Create a new course review'
      operationId: 4b2f32adda93339c1fb3127812a89101
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseReviewRequest'
      responses:
        201:
          description: 'New course review created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseReviewResource'
      security:
        -
          bearerAuth: []
  '/api/user/course-reviews/{id}':
    delete:
      tags:
        - CourseReviews
      summary: 'Delete a course review by ID'
      operationId: 55f6489f94ab36222229a253ad5aea45
      parameters:
        -
          name: id
          in: path
          description: 'ID of the course review'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Course review deleted'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    CourseReviewRequest:
      required:
        - course_id
        - rating
      properties:
        course_id:
          description: 'ID of the course to be reviewed'
          type: integer
          example: 1
        rating:
          description: 'Rating given to the course'
          type: number
          format: float
          example: 4.5
        content:
          description: 'Content of the course review'
          type: string
          example: 'This course was very informative and well-structured.'
      type: object
...
    StoreCourseRequest:
      title: 'Store Course Request'
      properties:
        category_ids:
          type: array
          items:
            type: integer
            example: '1'
        title:
          description: 'The title of the course.'
          type: string
          maxLength: 255
        description:
          description: 'The description of the course.'
          type: string
        price:
          description: 'The price of the course.'
          type: number
          format: float
        author:
          description: 'The author of the course.'
          type: string
          maxLength: 255
        thumbnail_image:
          description: 'Thumbnail image.'
          type: file
      type: object
...
    UpdateCourseRequest:
      title: 'Update Course Request'
      properties:
        title:
          description: 'The updated title of the course.'
          type: string
          maxLength: 255
        description:
          description: 'The updated description of the course.'
          type: string
        price:
          description: 'The updated price of the course.'
          type: number
          format: float
        author:
          description: 'The updated author of the course.'
          type: string
          maxLength: 255
        thumbnail_image:
          description: 'Thumbnail image.'
          type: file
      type: object
...
    CategoryResource:
      title: 'Category Resource'
      description: 'Represents a category resource.'
      properties:
        id:
          description: 'The unique identifier for the category.'
          type: integer
        parent_category:
          $ref: '#/components/schemas/CategoryResource'
        name:
          description: 'The name of the category.'
          type: string
        courses:
          description: 'List of courses associated with the category.'
          type: array
          items:
            $ref: '#/components/schemas/CourseResource'
        topics:
          description: 'List of topics associated with the category.'
          type: array
          items:
            $ref: '#/components/schemas/TopicResource'
      type: object
    ChapterResource:
      title: 'Chapter Resource'
      description: 'Represents a chapter resource.'
      properties:
        id:
          description: 'The unique identifier for the chapter.'
          type: integer
        course_id:
          description: 'The unique identifier of the course associated with the chapter.'
          type: integer
        title:
          description: 'The title of the chapter.'
          type: string
        lectures:
          description: 'A collection of lectures related to the chapter.'
          type: array
          items:
            $ref: '#/components/schemas/LectureResource'
      type: object
    CourseResource:
      title: 'Course Resource'
      description: 'Represents a course resource.'
      properties:
        id:
          description: 'The unique identifier for the course.'
          type: integer
        title:
          description: 'The title of the course.'
          type: string
        description:
          description: 'The description of the course.'
          type: string
        price:
          description: 'The price of the course.'
          type: number
        author:
          description: 'The author of the course.'
          type: string
        categories:
          description: 'List of categories associated with the course.'
          type: array
          items:
            $ref: '#/components/schemas/CategoryResource'
        reviews:
          description: 'List of reviews associated with the course.'
          type: array
          items:
            $ref: '#/components/schemas/CourseReviewResource'
      type: object
    CourseReviewResource:
      title: 'Course Review Resource'
      description: 'Represents a course review resource.'
      properties:
        id:
          description: 'The unique identifier for the review.'
          type: integer
        user:
          $ref: '#/components/schemas/UserResource'
        rating:
          description: 'The rating given in the review.'
          type: integer
        content:
          description: 'The content of the review.'
          type: string
        created_at:
          description: 'The timestamp when the review was created.'
          type: string
          format: date-time
      type: object
    LectureResource:
      title: 'Lecture Resource'
      description: 'Represents a lecture resource.'
      properties:
        id:
          description: 'The unique identifier for the lecture.'
          type: integer
        chapter_id:
          description: 'The unique identifier of the chapter associated with the lecture.'
          type: integer
        title:
          description: 'The title of the lecture.'
          type: string
        resources:
          description: 'A collection of resources related to the lecture.'
          type: array
          items:
            $ref: '#/components/schemas/LectureResourceResource'
        video:
          $ref: '#/components/schemas/LectureVideoResource'
        created_at:
          description: 'The date and time when the lecture was created.'
          type: string
          format: date-time
        updated_at:
          description: 'The date and time when the lecture was last updated.'
          type: string
          format: date-time
      type: object
    LectureResourceResource:
      title: LectureResourceModelResource
      description: 'Lecture resource model resource'
      properties:
        id:
          description: 'ID of the lecture resource'
          type: integer
        lecture_id:
          description: 'The ID of the lecture associated with this resource'
          type: integer
        title:
          description: 'Title of the lecture resource'
          type: string
        link:
          description: 'A link to the lecture resource'
          type: string
          format: uri
      type: object
      xml:
        name: LectureResourceModelResource
    LectureVideoResource:
      title: LectureVideoResource
      description: 'Lecture video resource'
      properties:
        data:
          properties:
            id:
              description: 'Unique identifier for the lecture video'
              type: integer
              example: 1
            url:
              description: 'URL of the video'
              type: string
              example: 'http://example.com/video.mp4'
            thumbnail_url:
              description: 'Thumbnail URL of the video'
              type: string
              example: 'http://example.com/thumbnail.jpg'
            duration:
              description: 'Duration of the video in seconds'
              type: integer
              example: 3600
            lecture_id:
              description: 'Foreign key to Lecture'
              type: integer
              example: 10
            created_at:
              description: 'Creation date of the video'
              type: string
              format: date-time
            updated_at:
              description: 'Last update date of the video'
              type: string
              format: date-time
          type: object
      type: object
      xml:
        name: LectureVideoResource
    TopicResource:
      title: 'Topic Resource'
      description: 'Represents a topic resource.'
      properties:
        id:
          description: 'The unique identifier for the topic.'
          type: integer
        name:
          description: 'The name of the topic.'
          type: string
        categories:
          description: 'List of categories associated with the topic.'
          type: array
          items:
            $ref: '#/components/schemas/CategoryResource'
      type: object
    UserResource:
      title: 'User Resource'
      description: 'Represents a user resource.'
      properties:
        id:
          description: 'The unique identifier for the user.'
          type: integer
        display_name:
          description: 'The display name of the user.'
          type: string
        email:
          description: 'The email address of the user.'
          type: string
          format: email
        created_at:
          description: 'The timestamp when the user was created.'
          type: string
          format: date-time
        updated_at:
          description: 'The timestamp when the user was last updated.'
          type: string
          format: date-time
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Login with email and password from user module to get the authentication token'
      name: Authorization
      in: header
      scheme: bearer
...